feat: BookmarkCoursesScreen 정렬 제거, 화면 크기 조정, UI 로직 수정 등

+++ app/src/main/java/com/choius323/saisai/data/course/remote/model/DeleteBookmarkCoursesDto.kt
@@ -4,5 +4,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class DeleteBookmarkCoursesDto(
-    val courseIds: List<Int>,
+    val courseIds: List<Long>,
 )
